Creation of index page with input for username AND development of home page
